Updating the intial documenation with some corrections and additional information in the README.md file
This commit contains a little correction on the contents in the README.md file, where we update the heading (header) of the file (which was written a bit improperly earlier).  Also we add up the "About the author" section to file, where we have written just two lines giving the information about the author's name and the date in which the script / project was started. But, more information will be added in the next few days.